Analysis Ticket #221,#256: Allow transcoding of sdat from genorepr to bit-packed representations without materialization.  Added shim functions to read text streams in ldat, sdat, and triple forms and output in bit-packed form. Added and improved many test cases.
                 And so it continues...

Welcome

       to

       the ongoing saga

                     of
                          BINARY

                    0101010101010101010
                    A.R.M.A.G.E.D.D.O.N
                    1010101010101010101

Consider these changes as the precuror to the second of four major
refactorings needed to merge binary support.  This next refactoring involves
the complete excision of the genoarray representation objects, since they
are no longer useful or needed.  Expect to see that land over the weekend or
Monday.  Fear not, because genoreprs will be replaced by string input and
output parsers, so we will retain the ability to read and write variously
delimited text genotype files.