Analysis Ticket #221,#256: Merging is getting fixed.  Sort of...
                          BINARY

                    0101010101010101010
                    A.R.M.A.G.E.D.D.O.N
                    1010101010101010101

Okay.  This is going to take some explaining.  Pretend for a moment that we
store genotypes for each locus using a different encoding.  When we have two
loci with the same name, we should ideally store them using the same
encoding.  The only question is when we make this distinction.

Strategy 1: Be purists and ensure that all loci with the same name always
have the same encoding.  This makes many lightweight operations, like
renaming loci, very expensive.  In fact, a lot of things get rather
complicated.  However, merging genotypes is a snap.  Binary writers and
packed formats currently assume Strategy 1.

Strategy 2: Be pragmatic and only merge locus models when it is inexpensive
to do so.  Pay the price only when merging and outputting binary streams.

I've begun to implement Strategy 1, however it is very expensive.  Maybe too
expensive.  Merging now works under Strategy 1 except for:
  filter_genomatrixstream_by_row (br0ken)
  merge_genomatrixstream_list    (br0ken and disabled -- falls back to triples)