Analysis Ticket #221: Add recoding to common genotype model to genomatrix streams and use it to fix merge_genomatrixstream_list.
The same solution can be used to remove most of the ugly from the other
matrix and triple stream merging code.  Expect to see those cleaned up next.

If you're paying attention, it means I've chosen Strategy 1, although that
means a bit more work when we rename, but a whole ton of savings and sanity
when we merge.  Thankfully most of it can be optimized away if we're clever.