Minor cleanups:
1. Models that cannot be fit will now trigger fallback to alternative
models.  This is more sensible than finding a theoretically good design
matrix that always blows up numerically.

2. Add very preliminary support for models with degenerate dependent
categories.  We don't quite get it right, since categories will not always
be output in the right places, though the headers will now align.  This is
an improvement over the current wrongier behavior.  I've left a large FIXME
to indicate that my job is never done.

3. Remove the partial printing of models that seem valid, but can't be fit.
It is now back to all-or-nothing.  I'll likely revisit this decision when I
start printing the actual model with each result row, since users find
reading the entrails, ie degrees of freedom, a bit arcane.