Major cleanup of genotype reader augmented filename parsing.  Defaults and processing of unique and genorepr are now handled by each reader.  This may seem like a step backward, since the old code may have seemed more unified. However, it is necessary if we are to move to a more pluggable IO framework.
The only internal API change is that genorepr must be passed as a kwarg to
the standard text formats.  Does not affect load_genostream, which already
had this requirement.

Only one minor bug was fixed -- genotriple streams were assumed to be unique
incorrectly.  The default may change back to assuming uniqueness, but that
has yet to be determined.

Upcoming cleanups:
  - add augmented filename parsing to writers (needed to support locus, map,
    and other metadata output files)
  - consider what to do with reader genome and phenome arguments