Locus models read from binary files should not be considered fixed, as that implies all alleles have been seen.
NB: This is really just moving the underlying problem elsewhere, but is a
reasonable compromise for the time being.  The better solution is to tighten
up semantics of fixed and ensure that no undue model sharing occurs.  This
is currently safe to do because model identity/compatibility is done via
object identity.  Should this ever be relaxed, then even greater care must
be taken.