Update GLU copyright and license as agreed upon in SAIC-Frederick Subcontract S07-041 modification 7, effective 2/5/2008.
New copyright and license text:

BSD-like open source license for distribution of GLU software:

Copyright (c) 2008, BioInformed LLC and the U.S. Department of Health &
Human Services. Funded by NCI under Contract N01-CO-12400.

BioInformed LLC and the U.S. Department of Health & Human Services
(?Copyright Holders?) hereby grant to the public a perpetual, irrevocable,
royalty-free non-exclusive, worldwide license to use, copy, distribute,
display and prepare derivatives of the software in source or binary forms,
with or without modification, subject only to the following conditions:

  o Redistributions of source code must retain the above copyright notice,
    this list of conditions and the following disclaimer.

  o Redistributions in binary form must reproduce the above copyright
    notice, this list of conditions and the following disclaimer in the
    documentation and/or other materials provided with the distribution.

  o Neither the name of BioInformed LLC nor the names of its contributors
    nor the U.S. Department of Health & Human Services may be used to endorse
    or promote products derived from this software without specific prior
    written permission.

THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
POSSIBILITY OF SUCH DAMAGE.