Add option to ignore genotype or phenotype metadata.  This is a stop-gap to a more general mechanism to override or discard such metadata when merging data from differently annotated sources.
The use-case that inspired this change was merging data typed on the
Illumina HumanHap 1M and 610Q chips.  They're rather different and are best
sorted out as a post-processing step.  For smaller datasets in the past I've
just converted to .sdat/.ldat to strip the metadata.  I've even hot-wired the
HDF5 version attributes to trick GLU into thinking the file was in V1 format
to ignore metadata.  However, those are both uglier solutions that this less
than ideal workaround.