Improve phenotype file parsing by parameterizing the pid and dependent variable columns.  Also use only phenotype columns used as covariates, since any missing values will supress observations.  It doesn't make sense to do this for columns not referenced in the model.
Parameterized the label for the "scan locus" in model, test, and display
formulae.  Model checking for was corrected to check for the presense or
absense of only the scan locus in the model and null-null, respectively.
The previous behavior tested for any loci, which served to disallow fixed
loci in null models and incorrectly validate non-null models with no
genotype scan terms.