Change fairly low level behavior or genotypes to compare equal and hash equal based on allelic comparison rather than object comparison.  This is yet another building-block to having dynamic locus representations.
To explain, this has to do with how models evolve within a given "encoding
scope" -- the scope where a genotype stream may be passed incrementally as
either objects or encoded vectors.

NOW:  Models are constant, contents may change.  Alleles and genotypes may
      be allocated on the fly, within limits set at construction time, but
      models objects for a locus remain constant.

SOON: Models change, contents are constant.  Alleles and genotypes are
      allocated to models at construction time, but model objects are
      replaced when new alleles are detected.

In the "NOW" model, it was convenient to have genotypes from different
models not compare equal or hash equal.  This was never a critical
assumption, but it was convenient.  Nowever, in the "SOON" model, a given
locus in SDAT stream may contain genotypes from several different models.
For example, the SDAT column [None,AA,AB,BB] will have 3 distinct models
(shown by allele lists): [], [A], [A,B],[A,B].