Add experimental support for multiple --includeloci and --excludeloci options.  Previously, only the last occurrence would be used and no warning was generated.  Now, the intersection of multiple includes and the union of multiple excludes is computed when more than one is specified.
Taking the intersection of includes may be a bit counter-intuitive, but the
logic is as follows: say one has a command like:

 > glu transform --includesamples=controls.lst infile.sbat -o outfile.sbat

and one would like to further restrict the input to only the subjects from a given study:

 > glu transform --includesamples=controls.lst --includesamples=study_A.lst [...]

is a natural construction.  If one wanted both controls and members if that
study, one could simply concatenate the lists together.  Obtaining the
intersection is not as simple.  If inclusion is the negation of exclusion,
it also argues for the intersection semantics by a natural application of
DeMorgan's laws (http://en.wikipedia.org/wiki/De_Morgan's_laws).

Union on exclusion is a must easier sell.  If one has multiple lists of
subjects to exclude, then by gum they should be gone!

This change was easy to make, since all (known) uses of this interface are
through the genolib.io.geno_options function.  Score one for good design!