Fix several minor issues in convert.from_lbd.  Sample includes and excludes now skip parsing records.  Pairs of genotypes and scores are now read properly, instead of mis-using the groupby itertool.  GC score filtering is now done directly on numpy arrays.
The result is that broken LBD files with dummy entries can not be ready with
the appropriate excludes.  Plus, the parser is now ~10% faster.  Further
speedups will require ridding ourselves of the csv parser and using a custom
one.

Added the 'chunk(iter,n)' generator to lib.utils that yields chunks (tuples) of
n items from iter.