Title: I cane speel reel gud
The GLU code base is large and often parts are written late at night, so
typos and spelling mistakes are a simple fact of life.  Typos in code are
rapidly ferreted out, but mistakes in non-imperative portions of the code
base may linger embarrassingly for years.  Thankfully I now have a tool to
scan for such landmines.  I've finally gotten around to implementing a
spell-checker module for Python source code strings and comments.  Please
try not to laugh too much at the corrections contained within this diff.