Add column filtering and value computation based on simple Python expressions in util.table, util.join, assoc.logit1, and assoc.linear1.  This support is fairly raw, since any errors are propagated back to the user.  It will likely be refined to produce warnings.
Here are some examples:

Subset a subject.def file for individuals with age between 20 and 29,
allowing for blanks as missing values::

  > glu util.table --filterexpr="20 <= int(fields[AGE_DX] or 0) < 30" subject.def

Create an AGE_SQUARED column::

  > glu util.table --columnexpr=AGE_SQUARED="int(fields[AGE_DX])**2 if fields[AGE_DX] else ''" subject.def

Documentation will be added later, but this will be a power-user feature
that few outside of NCI will likely use.  That is okay, since it is darn
useful for us and simple code to maintain.