Add glu.lib.pycompat module which contains the Counter and namedtuple collection objects from Python 2.7.
Counter replaces tally, though the tally function remains and simply
forwards calls to Counter.  tally will be depreciated and removed in a later
revision.  All callers of tally were updated to import and use Counter,
since this will be the official Python stdlib name of this object in Python
2.7.

namedtuple is not currently used, but the plan is to replace several
instances of awkward positional references in various corners of GLU with
namedtuples to enhance readability and robustness.